Added algorithm and Graph prototype in Java
This algorithm creates a graph with weighted edges and finds the
shortest distance between two given nodes. Hence we can use it to find
the shortest distance, then we change the graph, we find the shortest
distance again, and lastly we compare the two to see if there was
improvement.